requirements.txt
appdirs==1.4.4
argon2-cffi==20.1.0
attrs==20.1.0
backcall==0.2.0
black==19.10b0
bleach==3.1.5
Brotli==1.0.7
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
click==7.1.2
cycler==0.10.0
dash==1.14.0
dash-core-components==1.10.2
dash-cytoscape==0.2.0
dash-html-components==1.0.3
dash-renderer==1.6.0
dash-table==4.9.0
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.2
entrypoints==0.3
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
gevent==20.6.2
greenlet==0.4.16
idna==2.10
imageio==2.9.0
importlib-metadata==1.7.0
interpret==0.2.1
interpret-core==0.2.1
ipykernel==5.3.4
ipython==7.17.0
ipython-genutils==0.2.0
isort==5.4.2
itsdangerous==1.1.0
jedi==0.17.2
Jinja2==2.11.2
joblib==0.16.0
json5==0.9.5
jsonschema==3.2.0
jupyter-client==6.1.6
jupyter-core==4.6.3
jupyterlab==2.2.5
jupyterlab-server==1.2.0
kiwisolver==1.2.0
lime==0.2.0.1
MarkupSafe==1.1.1
matplotlib==3.3.1
mistune==0.8.4
msgpack==1.0.0
nbconvert==5.6.1
nbformat==5.0.7
neovim==0.3.1
networkx==2.4
notebook==6.1.3
numpy==1.19.1
packaging==20.4
pandas==1.1.1
pandocfilters==1.4.2
parso==0.7.1
pathspec==0.8.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.2.0
plotly==4.9.0
prometheus-client==0.8.0
prompt-toolkit==3.0.6
psutil==5.7.2
psycopg2==2.8.5
ptyprocess==0.6.0
pycairo==1.19.1
pycparser==2.20
Pygments==2.6.1
pynvim==0.4.1
pyparsing==2.4.7
pyrsistent==0.16.0
python-dateutil==2.8.1
python-igraph==0.8.2
pytz==2020.1
PyWavelets==1.1.1
pyzmq==19.0.2
regex==2020.7.14
requests==2.24.0
retrying==1.3.3
RuleFit @ git+git://github.com/christophM/rulefit.git@b1657af4b41df59e2ae64bb1767dbaf5ff1ed7fe
SALib==1.3.11
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.10.1
Send2Trash==1.5.0
shap==0.35.0
six==1.15.0
SQLAlchemy==1.3.19
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
threadpoolctl==2.1.0
tifffile==2020.8.13
toml==0.10.1
tornado==6.0.4
tqdm==4.48.2
traitlets==4.3.3
treeinterpreter==0.2.2
typed-ast==1.4.1
urllib3==1.25.10
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
